requirements.txt
autokeras==1.0.2
tensorflow==2.1.0
tpot==0.11.1
h2o==3.28.1.2
statsmodels==0.11.1
pandas==1.0.1
dataclasses
ete3==3.1.1
numpy==1.18.0
pytest==5.2.0
anytree==2.8.0
typing==3.7.4.1
scikit_learn==0.22.1
matplotlib==3.0.2
networkx==2.4
xgboost==1.0.1
Pillow==7.0.0
imageio==2.8.0
pydot==1.4.1
keras==2.3.1